Add Test Action used only in branches
This Test Action will just run acceptance and unit tests, and run
exclusively on `push` and `pull_request` events on branches.

This has been done because I'll have a separate action that runs only
on `push` events on **tags** for release.

This commit also removes an unneeded step from the test action.
Dependencies are installed automatically when running tests.