Add release action
This action will run only on `push` events with tags matching the `v*`
filter. This action will still run tests like the test action, but it
will also go ahead with the release process: mainly building the
provider with the proper tag, and packing and releasing the provider
in Github.