Document using binary from Release Action
Currently I've set up a Release Action using `ubuntu-latest`, which
will build the provider binary on any tag push. This allows me to
automate building and pushing releases with a `linux_amd64` binary.

But the README didn't mention this, and seemed like the only option to
use this provider was to build from source, which might not always be
the case.