Force changes in domain arg to re-create resource
The domain argument is somewhat special, since it doesn't represent a
value stored with the record, but rather it's used for the API to know
to what domain the record you want to create should be attached to.

Which means that changing this argument just means the API will create
the record under a new domain, so it makes sense that if you change
this value in your Terraform config, it should destroy the previous
domain's record, and re-create it under the new one.