Test workflow: Test in Linux, Windows and MacOS
This is done using a matrix which has a max-parallel of 1. Increasing
this number would surely lower the CI run time, but since the
acceptance tests aren't isolated and they create and later on destroy
the created infrastructure, we need to make sure there's not two test
suites creating and checking resources at the same time which would
lead to race conditions and failing tests.