Release workflow: Build binaries for all major OS
I've set up a combination of multiple jobs, some matrix jobs that run
in all three major OS (Linux, Windows, and MacOS).

First, acceptance and unit tests are run in all the OS. This job
matrix does not run in parallel since acceptance tests are not
isolated and they depend on a common resource/infrastructure. Which
means running the same tests on different platforms at once would lead
to race conditions and therefore failed tests.

Second, at the same time the tests are running, there's a job called
`version` running just on Ubuntu in charge of some initial set up
tasks. These are:
* Parsing the version out of the `github.ref` variable, which will
  hold something like `refs/tags/v1.2.3`. For multiple tasks later on
  we need to fetch just the version variable (e.g. `v1.2.3`).
* A release is also created and marked as draft. This is because
  pushing a published release (not draft) will send out a notification
  with the assets it has attached. If you asynchronously attach new
  assets later on, these will not be reflected in the notification.

Third, the build job depends on the test and version jobs running and
finishing successfully. It sets up another matrix of Linux, Windows
and MacOS, and builds the Golang provider binary, with the associated
version number, in each platform. These platforms are only 64bit, but
that's good enough for our purposes.

It also makes use of a community action to `tar` the Golang provider
binary along with the README.md and LICENSE. Normally I would have
just used the `tar` CLI tool, but I don't believe that is available
under Windows, which we are running in this job.

Once the `tar.gz` file for any OS is built, they get uploaded to the
previously drafted release.

Fourth and final, we run a job called publish, which depends on the
version and build jobs, which just makes use of the `hub` CLI tool to
programmatically, through the Linux shell, publish the previously
created draft release, which will trigger the release notification but
this time will include all the assets at once (rather than being
asynchronously added later on at some point).