Remove pull_request trigger on Test workflow
It doesn't work as it expected it to. I thought that for PRs, and
running a test suite on PRs, either internal or from forked
repositories, I'd have to use the `pull_request` event.

However, there's an issue with this: right now there's no easy way to
configure the `pull_request` event so it only runs on _external_ pull
requests (as in, PRs from forks).

Which means that internal PRs from tools such as Dependabot will
trigger two CIs, one for the `push` event, and one for the
`pull_request` event.

So I've decided to just remove the `pull_request` event, since it will
fail on external PRs anyway since secrets are not shared.