Add new Terraform Registry style documentation
Once Terraform v0.13 is released we'll be able to finally ship
third-party providers without relying on the user to compile and place
the binaries in specific directories.

v0.13 is still in beta but it should be released at some point in
early August if all works well.

For that reason, I decided to put in the work already to follow the
documentation practices for the new Registry, which will fetch files
from the `docs` directory and render/display them on the Registry
website.