Use GoReleaser in Release workflow
The Terraform documentation recommends using this tool to manage
cross-compilation, archiving, and generating the checksums of the
artifacts.

It can also manage signing, but for now I'll keep doing that manually.
Any tag will create a draft release, and from there I'll download the
checksum file, sign it manually with `gpg --detach-sign` and then
upload it back to the release assets and publish it.