Document provider now in the Terraform Registry
With the inclusion of a Terraform Registry, we no longer have to
manually compile and install the provider binaries in specific
locations in our local machines. Terraform v0.13 has a new syntax to
fetch providers from the Registry and download them, abstracting the
user from knowing how to compile and where to save the provider
binaries to use them.

And so with this change, the `README.md` has been updated to reflect
the automatically rendered online documentation in the Terraform
Registry website, the link to the provider in the Registry, and a
couple other things.

The Action for sending a dispatch to the
`terraform-provider-njalla-pkgbuild` repository to build an Arch Linux
package of this provider has been removed, since that project is
deprecated, now that the Registry exists and we no longer need to
manually install third-party providers.