Fix records string IDs
The Njalla API has had a silent change where they've started using
string IDs for records instead of the previous integer IDs. This
required changes to the `gonjalla` package, and also requires changes
to the provider configuration to no longer expect/convert from `int`
to `string` or the other way around all over, since now the IDs are
strings, same as the Terraform IDs for resources are strings.